third_party/WebKit/LayoutTests/fast/html/imports/import-style-basic.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<head>
<link id="importLink" rel="import" href="resources/style-basic-child.html">
<style>
.red-from-import {
    color: white; // Overriding with white.
};
</style>
</head>
<body>
<h1>These elements should be styled appropriately:</h1>
<div id="testContainer">
  <div id="shouldBeBlue" class="blue-from-import">Should be blue.</div>
  <div id="shouldBeWhite" class="red-from-import">Should be white.</div>
  <div id="shouldBeYellow" class="yellow-from-import">Should be yellow.</div>
  <div id="shouldBeTeal" class="gray-from-subimport">Should be teal.</div>
  <div id="shouldBeGreen" class="green-from-import">Should become green.</div>
  <div id="shouldBeAqua" class="aqua-from-import">Should become aqua.</div>
</div>

<script>
window.jsTestIsAsync = true;

// Master's CSSOM contains master's sheets only.
shouldBe("document.styleSheets.length", "2"); // Extra one comes from js-test.js
// Import's CSS should have its own
shouldBe("importLink.import.styleSheets.length", "1");

// a trivial case
shouldBeEqualToString("window.getComputedStyle(shouldBeBlue).color", "rgb(0, 0, 255)");
// Nested case
shouldBeEqualToString("window.getComputedStyle(shouldBeYellow).color", "rgb(255, 255, 0)");
// Overriding import from the master
shouldBeEqualToString("window.getComputedStyle(shouldBeWhite).color", "rgb(255, 255, 255)");
// Overriding subimport from parent import
shouldBeEqualToString("window.getComputedStyle(shouldBeTeal).color", "rgb(0, 128, 128)");

// Dynamically adding
shouldBeEqualToString("window.getComputedStyle(shouldBeGreen).color", "rgb(0, 0, 0)");
var styleWithGreen = importLink.import.createElement("style");
styleWithGreen.innerHTML = ".green-from-import { color: green; }";
importLink.import.head.appendChild(styleWithGreen);
shouldBeEqualToString("window.getComputedStyle(shouldBeGreen).color", "rgb(0, 128, 0)");
shouldBe("importLink.import.styleSheets.length", "2");

// Dyanically removing
Array.prototype.forEach.call(importLink.import.querySelectorAll("style"), function(e) { e.remove(); });
shouldBeEqualToString("window.getComputedStyle(shouldBeBlue).color", "rgb(0, 0, 0)");
shouldBeEqualToString("window.getComputedStyle(shouldBeGreen).color", "rgb(0, 0, 0)");
shouldBe("importLink.import.styleSheets.length", "0");

// Dymically adding an import
var dynamicLink = document.createElement("link");
dynamicLink.setAttribute("rel", "import");
dynamicLink.id = "importLinkDynamic";
dynamicLink.setAttribute("href", "resources/style-basic-child-dynamic.html");
dynamicLink.onload = function() {
    shouldBeEqualToString("window.getComputedStyle(shouldBeAqua).color", "rgb(0, 255, 255)");
    shouldBe("importLinkDynamic.import.styleSheets.length", "1");
    finishJSTest();
};

document.head.appendChild(dynamicLink);

</script>
</body>
</html>
