third_party/WebKit/LayoutTests/fast/html/imports/import-style-basic-expected.txt

CONSOLE WARNING: Styling master document from stylesheets defined in HTML Imports is deprecated. Please refer to https://goo.gl/EGXzpw for possible migration paths.
PASS document.styleSheets.length is 2
PASS importLink.import.styleSheets.length is 1
PASS window.getComputedStyle(shouldBeBlue).color is "rgb(0, 0, 255)"
PASS window.getComputedStyle(shouldBeYellow).color is "rgb(255, 255, 0)"
PASS window.getComputedStyle(shouldBeWhite).color is "rgb(255, 255, 255)"
PASS window.getComputedStyle(shouldBeTeal).color is "rgb(0, 128, 128)"
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 0, 0)"
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 128, 0)"
PASS importLink.import.styleSheets.length is 2
PASS window.getComputedStyle(shouldBeBlue).color is "rgb(0, 0, 0)"
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 0, 0)"
PASS importLink.import.styleSheets.length is 0
PASS window.getComputedStyle(shouldBeAqua).color is "rgb(0, 255, 255)"
PASS importLinkDynamic.import.styleSheets.length is 1
PASS successfullyParsed is true

TEST COMPLETE
These elements should be styled appropriately:
Should be blue.
Should be white.
Should be yellow.
Should be teal.
Should become green.
Should become aqua.
